refactor(layout): move hooks and compose helpers out of verbs/
verbs/ should contain only CLI verbs. The previous layout mixed in:
- four *_hooks.py files (kernel-side hook registration, not verbs)
- one _compose_gen.py (compose-generation glue, used by up + upgrade)

Moves (git mv preserves history):
- verbs/{up,build,lifecycle,snapshot}_hooks.py → hooks/{up,build,lifecycle,snapshot}.py
- verbs/_compose_gen.py → compose/generators.py

Also drops a layer-violating re-export in core/orchestrator.py
(`from sanity_gravity.hooks.up import register_builtin_up_hooks`) that
nothing actually consumed; the verb modules now import their hooks
directly from `sanity_gravity.hooks.<verb>`.

Imports updated across:
- sanity_gravity/verbs/{up,build,lifecycle,snapshot,upgrade}.py
- sanity_gravity/core/orchestrator.py docstring
- 7 test files

The new top-level layout has clearer concept-per-package:
  hooks/    - builtin verb-lifecycle hook implementations
  verbs/    - CLI verb wrappers (build context, dispatch to orchestrator)
  compose/  - YAML primitives (builder) + generators
  core/     - kernel: orchestrator + eventbus + reporter + command

All 209 unit tests pass.
